+++ src/app/beer_garden/db/mongo/util.py
@@ -126,22 +126,27 @@ def prune_topics():
         None
     """
 
-    from .models import Garden, System, Topic
+    from .models import System, Topic
 
     command_hash = []
 
-    for garden in Garden.objects():
-        if garden.name == config.get("garden.name"):
-            garden.systems = System.objects(local=True)
-        for system in garden.systems:
-            for instance in system.instances:
-                for command in system.commands:
-                    command_hash.append(
-                        (
-                            f"{garden.name}.{system.namespace}.{system.name}."
-                            f"{system.version}.{instance.name}.{command.name}"
-                        )
+    for system in System.objects().only(
+        "garden_name",
+        "name",
+        "namespace",
+        "version",
+        "instances.name",
+        "commands.name",
+    ):
+
+        for instance in system.instances:
+            for command in system.commands:
+                command_hash.append(
+                    (
+                        f"{system.garden_name}.{system.namespace}.{system.name}."
+                        f"{system.version}.{instance.name}.{command.name}"
                     )
+                )
 
     deleted_topic_count = 0
     deleted_subscriber_count = 0