+++ CHANGELOG.md
@@ -28,6 +28,8 @@ TBD
 - UI code is now unminimized with source map to enable enhanced debugging of released code (#1942)
 - Improved None checking for Request Parents
 - Updated internal class from EventProcessor to ReplicationProcessor for improved clarity (#1780)
+- Added Support to track Garden Name on System Models (#1950)
+
 - Updated latest sytem version to prioritize running but still return a system if not (#1949)
 - Updated Jobs pages to trigger off internal Job events for status changes (#1878, #1951)
 - Updated Publish Request to ensure topic data is added to Request Metadata (#1954)